+++ src/components/link/arrow-link.tsx
@@ -1,6 +1,7 @@
 import { IndexArrowIcon } from '@/components/icons/index-arrow-icon'
 import { type ILinkProps } from '@interfaces/link-props'
 import { cn } from '@lib/shadcn-utils'
+import { useState } from 'react'
 import { BaseLink } from './base-link'
 
 export function ArrowLink({
@@ -9,14 +10,20 @@ export function ArrowLink({
   children,
   ...props
 }: ILinkProps & { stroke?: string }) {
+  const [hover, setHover] = useState<'initial' | 'on' | 'off'>('initial')
+
   return (
     <BaseLink
       className={cn('inline-flex flex-row items-center group', className)}
+      onMouseEnter={() => {
+        setHover('on')
+      }}
+      onMouseLeave={() => setHover('off')}
       {...props}
     >
       {children}
 
-      <IndexArrowIcon className="w-4" stroke={stroke} />
+      <IndexArrowIcon className="w-4" stroke={stroke} hover={hover} />
     </BaseLink>
   )
 }