+++ src/components/posts/latest-post.tsx
@@ -1,12 +1,8 @@
-import type { IPostProps } from "./post-tags"
-import { PreviewPost } from "./preview-post"
+import { HeroPost } from './hero-post'
+import type { IPostProps } from './post-tags'
 
 export function LatestPost({ post }: IPostProps) {
   return (
-    <PreviewPost
-      post={post}
-      showAvatarImage={false}
-      headerClassName="text-3xl"
-    />
+    <HeroPost post={post} showAvatarImage={false} headerClassName="text-3xl" />
   )
 }