+++ src/components/posts/rest-posts.tsx
@@ -2,22 +2,26 @@ import { cn } from '@/lib/shadcn-utils'
 import { HeroPostSmall } from './hero-post-small'
 import type { IPostsProps } from './hero-posts'
 
-export function RestPosts({ posts }: IPostsProps) {
+export function BaseRestPosts({ posts, showSectionLinks }: IPostsProps) {
   return (
-    <section>
-      <ul className="grid grid-cols-1 gap-x-12 gap-y-6 md:grid-cols-2 lg:grid-cols-3">
-        {posts.map((post, index) => (
-          <li key={index}>
-            <HeroPostSmall
-              post={post}
-              className={cn(
-                'pb-6 border-b border-border',
-                index < 3 && 'pt-6 border-t border-border'
-              )}
-            />
-          </li>
-        ))}
-      </ul>
+    <ul className="grid grid-cols-1 gap-x-12 gap-y-6 md:grid-cols-2 lg:grid-cols-3">
+      {posts.map((post, index) => (
+        <li key={index}>
+          <HeroPostSmall
+            post={post}
+            className={cn(index > 2 && 'pt-6 border-t border-border')}
+            showSectionLinks={showSectionLinks}
+          />
+        </li>
+      ))}
+    </ul>
+  )
+}
+
+export function RestPosts({ posts, showSectionLinks }: IPostsProps) {
+  return (
+    <section id="rest-posts" className="mb-24">
+      <BaseRestPosts posts={posts} showSectionLinks={showSectionLinks} />
     </section>
   )
 }